[Fix:ejr] 디렉토리 구조 재정의
