--- src/main/java/net/fullstack7/aaa/config/config
@@ -1,1 +1,0 @@
-config
