+++ src/main/java/net/fullstack7/aaa/controller/MemberController.java
@@ -1,9 +1,9 @@
-package net.fullstack7.controller;
+package net.fullstack7.aaa.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
-import net.fullstack7.dto.MemberDTO;
-import net.fullstack7.service.MemberServiceImpl;
+import net.fullstack7.aaa.dto.member.MemberDTO;
+import net.fullstack7.aaa.service.member.MemberServiceImpl;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
