--- src/main/java/net/fullstack7/aaa/controller/controller
@@ -1,1 +1,0 @@
-controller
