--- src/main/java/net/fullstack7/aaa/repository/repository
@@ -1,1 +1,0 @@
-repository
