+++ src/main/java/net/fullstack7/aaa/service/member/MemberServiceIf.java
@@ -1,0 +1,7 @@
+package net.fullstack7.aaa.service.member;
+
+import net.fullstack7.aaa.dto.member.MemberDTO;
+
+public interface MemberServiceIf {
+    void signUp(MemberDTO memberDTO);
+}
