--- src/main/java/net/fullstack7/aaa/service/service
@@ -1,1 +1,0 @@
-service
