--- src/main/java/net/fullstack7/aaa/util/util
@@ -1,1 +1,0 @@
-util
