--- src/main/java/net/fullstack7/service/MemberServiceIf.java
@@ -1,7 +1,0 @@
-package net.fullstack7.service;
-
-import net.fullstack7.dto.MemberDTO;
-
-public interface MemberServiceIf {
-    void signUp(MemberDTO memberDTO);
-}
